requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.1
joblib==1.5.1
kiwisolver==1.4.8
lxml==5.4.0
matplotlib==3.10.3
narwhals==1.41.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
plotly==6.1.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
xgboost==3.0.2
